- fixed quest menu icons: icon = 2 for available quests, else icon = 4!

+++ Services/WCell.RealmServer/Global/Map.cs
@@ -955,7 +955,7 @@ namespace WCell.RealmServer.Global
 			for (var i = 0; i < objs.Length; i++)
 			{
 				var obj = objs[i];
-				if (!(obj is Character) && !obj.IsOwnedByPlayer)
+				if (!(obj is Character) && !obj.IsOwnedByPlayer && !obj.IsDeleted)
 				{
 					// only delete things that are not Characters or belong to Characters
 					obj.DeleteNow();
